refactor(memory): extract reader and file-list controls
