+++ src/components/familiars-memory-view-detail.test.ts
@@ -4,6 +4,7 @@ import { readFile } from "node:fs/promises";
 
 const source = [
   await readFile(new URL("./familiars-memory-view.tsx", import.meta.url), "utf8"),
+  await readFile(new URL("./familiars-memory-files.tsx", import.meta.url), "utf8"),
   await readFile(new URL("./familiars-memory-utils.ts", import.meta.url), "utf8"),
 ].join("\n");
 const hook = await readFile(new URL("../lib/use-memory-file.ts", import.meta.url), "utf8");
