+++ src/components/familiars-memory-view-management.test.ts
@@ -1,6 +1,9 @@
 import assert from "node:assert/strict";
 import { readFile } from "node:fs/promises";
-const source = await readFile(new URL("./familiars-memory-view.tsx", import.meta.url), "utf8");
+const source = [
+  await readFile(new URL("./familiars-memory-view.tsx", import.meta.url), "utf8"),
+  await readFile(new URL("./familiars-memory-files.tsx", import.meta.url), "utf8"),
+].join("\n");
 
 assert.match(source, /useUndoDelete/, "uses the shared undo-delete hook");
 assert.match(source, /<UndoToast/, "renders the shared undo toast");
